fix: cleanup sw
